Separate utility functions into their respective static classes
