+++ Assets/Scripts/Utils/GeneralUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d03a993ee67f144ba6e06bf6ee26012
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
